+++ util/benchmarksUtil.js
@@ -1,7 +1,3 @@
-const utility = require('./utility');
-const config = require('../config');
-const redis = require('../store/redis');
-
 const benchmarks = {
   gold_per_min(m, p) {
     return p.gold_per_min;
@@ -24,9 +20,6 @@ const benchmarks = {
   tower_damage(m, p) {
     return p.tower_damage;
   },
-};
-
-const parsedBenchmarks = {
   stuns(m, p) {
     return p.stuns;
   },
@@ -35,36 +28,7 @@ const parsedBenchmarks = {
   },
 };
 
-function updateBenchmarks(match, cb, parsed) {
-  for (let i = 0; i < match.players.length; i += 1) {
-    const p = match.players[i];
-    const bms = parsed ? parsedBenchmarks : benchmarks;
-    // only do if all players have heroes
-    if (p.hero_id) {
-      Object.keys(bms).forEach((key) => {
-        const metric = bms[key](match, p);
-        if (metric !== undefined && metric !== null && !Number.isNaN(Number(metric))) {
-          const rkey = [
-            'benchmarks',
-            utility.getStartOfBlockMinutes(config.BENCHMARK_RETENTION_MINUTES, 0),
-            key,
-            p.hero_id,
-          ].join(':');
-          redis.zadd(rkey, metric, match.match_id);
-          // expire at time two epochs later (after prev/current cycle)
-          const expiretime = utility.getStartOfBlockMinutes(config.BENCHMARK_RETENTION_MINUTES, 2);
-          redis.expireat(rkey, expiretime);
-        }
-      });
-    }
-  }
-  return cb();
-}
 
 module.exports = {
-  benchmarks: {
-    ...benchmarks,
-    ...parsedBenchmarks,
-  },
-  updateBenchmarks,
+  benchmarks,
 };
